Refactor feature extraction and preprocessing: optimize band power and PSD calculations; add error handling for subject loading

+++ src/transforms/pca.py
@@ -111,4 +111,4 @@ class MyPCA(BaseEstimator, TransformerMixin):
             Transformed data of shape (n_samples, n_components)
         """
         X_centered = X - self.mean_
-        return np.dot(X_centered, self.components_)
+        return np.dot(X_centered, self.components_).astype(np.float64)
